- compass working now - 2 different compass approaches, using different sensors

+++ app/src/main/java/de/htwberlin/f4/ai/ba/coordinates/android/sensors/compass/CompassListener.java
@@ -6,5 +6,5 @@ import de.htwberlin.f4.ai.ba.coordinates.android.sensors.SensorListener;
  * Created by benni on 19.07.2017.
  */
 
-public interface CompassListener extends SensorListener<Float> {
+public interface CompassListener extends SensorListener<Integer> {
 }
